Allow config of PUID / PGID per OPTIONS.md
OPTIONS.md indicates that the user / group IDs of the sagesv user can be
configured with PUID / PGID environment variables to the container. This
wasn't working for me. After looking at the Dockerfile, the user ID
appears hard coded to 911 and the GID is just the next available id,
This changes the Dockerfile to match the description in OPTIONS.md. It
also changes the defaults to match OPTIONS.md (99 for UID and 100 for
GID).