Make PUID / PGID environment options actually work
The PUIG / PGID environment variables specified in OPTIONS.md didn't do
anything (weren't used in the Dockerfile). Updated Dockerfile to use
them. Also, changed docs to reflect default values of 911 for both. 99
was ok for PUID, but the Dockerfile was actually using 99. And, 100
wasn't valid (already in use) so made it 911 as well.